refactor(providers): extract PipelineRequestBodyEditor + PipelinePolicyTestHarness
Three production PipelinePolicy implementations had the same 8-line
"stream -> parse -> mutate JsonObject -> bail if not an object ->
serialize back as BinaryContent" sequence:

- OpenRouterReasoningExcludePolicy
- OpenAiCodexRequestPolicy
- VeniceAiSystemPromptOverridePolicy (in flight on PR #1138)

Extract the boilerplate into PipelineRequestBodyEditor.EditJsonBody so
each policy expresses only its mutation as an Action<JsonObject>. The
mutation logic is unchanged in every case; only the parse/serialize
plumbing is consolidated.

Same treatment for the test side: CapturePolicy + ProcessSync +
CreateMessage(JsonObject?) were duplicated between
OpenRouterReasoningExcludePolicyTests and
VeniceAiSystemPromptOverridePolicyTests. Hoist into
PipelinePolicyTestHarness so future policy tests get the harness for
free.

Net line count: roughly flat (+28). The win is having one place that
owns the parse/serialize pattern, not raw line reduction. The
individual policies are noticeably easier to skim:

- OpenRouterReasoningExcludePolicy: 57 -> 35 lines (-22)
- OpenAiCodexRequestPolicy: 150 -> 122 lines (-28)
- OpenRouterReasoningExcludePolicyTests: 149 -> 80 lines (-69)

PR #1138 will rebase on top of this and port the Venice policy + tests
to use the helper. Resolves #1137.

Verified:
- dotnet build (0 warnings, 0 errors)
- 134/134 Netclaw.Daemon.Tests Providers suite passes
- 20/20 OpenAiCodex tests pass (the highest-risk site)
- Add-FileHeaders.ps1 -Verify clean
- dotnet slopwatch analyze: 0 issues
