+++ src/Netclaw.Daemon.Tests/Providers/OpenRouterReasoningExcludePolicyTests.cs
@@ -1,11 +1,8 @@
-﻿// -----------------------------------------------------------------------
+// -----------------------------------------------------------------------
 // <copyright file="OpenRouterReasoningExcludePolicyTests.cs" company="Petabridge, LLC">
 //      Copyright (C) 2026 - 2026 Petabridge, LLC <https://petabridge.com>
 // </copyright>
 // -----------------------------------------------------------------------
-using System.ClientModel;
-using System.ClientModel.Primitives;
-using System.Text.Json;
 using System.Text.Json.Nodes;
 using Netclaw.Providers.OpenRouter;
 using Xunit;
@@ -24,7 +21,7 @@ public sealed class OpenRouterReasoningExcludePolicyTests
             ["messages"] = new JsonArray(new JsonObject { ["role"] = "user", ["content"] = "hello" })
         };
 
-        var result = ProcessSync(policy, body);
+        var result = PipelinePolicyTestHarness.RunSync(policy, body);
 
         Assert.NotNull(result);
         Assert.True(result!["reasoning"]?["exclude"]?.GetValue<bool>());
@@ -42,7 +39,7 @@ public sealed class OpenRouterReasoningExcludePolicyTests
             ["messages"] = new JsonArray(new JsonObject { ["role"] = "user", ["content"] = "think hard" })
         };
 
-        var result = ProcessSync(policy, body);
+        var result = PipelinePolicyTestHarness.RunSync(policy, body);
 
         Assert.NotNull(result);
         Assert.Equal("deepseek/deepseek-r1", result!["model"]?.GetValue<string>());
@@ -61,7 +58,7 @@ public sealed class OpenRouterReasoningExcludePolicyTests
             ["reasoning"] = new JsonObject { ["effort"] = "high" }
         };
 
-        var result = ProcessSync(policy, body);
+        var result = PipelinePolicyTestHarness.RunSync(policy, body);
 
         Assert.NotNull(result);
         // The policy overwrites any existing reasoning config
@@ -72,78 +69,12 @@ public sealed class OpenRouterReasoningExcludePolicyTests
     public void NoOps_WhenContentIsNull()
     {
         var policy = new OpenRouterReasoningExcludePolicy();
-        // Pipeline wrapper that captures the message without content
-        var pipeline = new CapturePolicy();
-        var message = CreateMessage(content: null);
+        var capture = new PipelinePolicyTestHarness.CapturePolicy();
+        var message = PipelinePolicyTestHarness.CreateMessage(null);
 
-        policy.Process(message, [policy, pipeline], 0);
+        policy.Process(message, [policy, capture], 0);
 
-        // Should pass through without crashing
-        Assert.True(pipeline.WasCalled);
+        Assert.True(capture.WasCalled);
         Assert.Null(message.Request.Content);
-    }
-
-    /// <summary>
-    /// Runs the policy synchronously and returns the modified JSON body.
-    /// </summary>
-    private static JsonObject? ProcessSync(OpenRouterReasoningExcludePolicy policy, JsonObject body)
-    {
-        var pipeline = new CapturePolicy();
-        var message = CreateMessage(body);
-
-        policy.Process(message, [policy, pipeline], 0);
-
-        Assert.True(pipeline.WasCalled, "Policy must call ProcessNext");
-
-        if (message.Request.Content is null)
-            return null;
-
-        using var stream = new MemoryStream();
-        message.Request.Content.WriteTo(stream, default);
-        return JsonSerializer.Deserialize<JsonObject>(stream.ToArray());
-    }
-
-    private static PipelineMessage CreateMessage(JsonObject? body)
-    {
-        var pipeline = ClientPipeline.Create();
-        var message = pipeline.CreateMessage();
-
-        if (body is not null)
-        {
-            var bytes = JsonSerializer.SerializeToUtf8Bytes(body);
-            message.Request.Content = BinaryContent.Create(BinaryData.FromBytes(bytes));
-        }
-
-        return message;
-    }
-
-    private static PipelineMessage CreateMessage(BinaryContent? content)
-    {
-        var pipeline = ClientPipeline.Create();
-        var message = pipeline.CreateMessage();
-        if (content is not null)
-            message.Request.Content = content;
-        return message;
-    }
-
-    /// <summary>
-    /// Terminal policy that records it was reached.
-    /// </summary>
-    private sealed class CapturePolicy : PipelinePolicy
-    {
-        public bool WasCalled { get; private set; }
-
-        public override void Process(
-            PipelineMessage message, IReadOnlyList<PipelinePolicy> pipeline, int currentIndex)
-        {
-            WasCalled = true;
-        }
-
-        public override ValueTask ProcessAsync(
-            PipelineMessage message, IReadOnlyList<PipelinePolicy> pipeline, int currentIndex)
-        {
-            WasCalled = true;
-            return default;
-        }
     }
 }
